【fix】 1.修改mybatis插件 2.调整配置文件 3.升级mybatis版本

+++ business/src/main/java/com/example/cloud/data/response/blog/BlogUserInfoResponse.java
@@ -25,15 +25,15 @@ public class BlogUserInfoResponse {
     /**
      * blog统计
      */
-    Integer blogCount;
+    Long blogCount;
 
     /**
      * 评论统计
      */
-    Integer commentCount;
+    Long commentCount;
 
     /**
      * 粉丝数
      */
-    Integer followCount;
+    Long followCount;
 }
